# perf: only cache docs that are fetched from cache