As per current implementation whenever `get_doc` is called, document is cached. However, this cache is only ever used by `get_cached_doc`. Going through the codebase of both FF/ERPNext you'll find that `get_doc` is used a lot more than `get_cached_doc`. So in many places, all this caching overhead is unnecessary.

Pros: faster `get_doc`, slightly lower memory usage. Correctness i.e. caching only what gets used from cache. 
Con: After this change. First call to `get_cached_doc` will always be a cache miss.

```
# Before
%timeit frappe.get_doc("Sales Order", "SAL-ORD-2021-00009")
22.1 ms ± 195 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

# After
%timeit frappe.get_doc("Sales Order", "SAL-ORD-2021-00009")
16.9 ms ± 18.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```


---

PS: Any better ideas for benchmarking are welcome.