@leela 2021-07-02
@ankush , Are you suggesting to use get_cached_doc instead of get_doc?

Ideally there should be only one API and that is a get_doc.(It should internally take care about caching). I assume there are two implementations because cache clearing is not happening the way it should happen.

We currently use get_cached_doc in the system only after get_doc is called at least once in a transaction (This is an untold rule). With this change, get_cached_doc may return old cached data because of cache clearing implementation is not reliable.