@ankush 2021-07-02
> @ankush , Are you suggesting to use get_cached_doc instead of get_doc?

I am not suggesting any change in how we structure code.  This is just a proposal for internal implementation change. `get_doc` is caching unnecessarily documents that are never fetched back from the cache.

> We currently use get_cached_doc in the system only after get_doc is called at least once in a transaction **(This is an untold rule)**. With this change, get_cached_doc may return old cached data because of cache clearing implementation is not reliable.

Ah... wasn't aware of this. Then again, using cache should imply that we won't be 100% sure about the correctness of data. Assuming it's done for perf reasons in cases where docs aren't changing that fast (like global settings, non-transactional data). A decent fix could be to have some eviction policy based on time or MRU of sorts?


`db.get_value(cache=True)` has this correctnss problem too (probably worse though, as it implements a separate cache for values).